requirements.txt
django
djangorestframework
mysqlclient
Pillow
drf-yasg